Refactor: Split five oversized modules
gerrit/service and gerrit/submit_manager keep their module identity and
shed mixins, because build_client and create_url_builder are patched in
their own namespaces at 88 test sites.  copilot_handler,
resolve_conflicts and progress_tracker become packages.

progress_tracker turned out to be a monkeypatch target too, which the
brief for this work had missed: thirteen tests swap rich's Live class
on the module.  A call site resolves names from its own module globals,
so once ProgressTracker moved into a submodule, patching the package
attribute silently stopped reaching it.  The tests now patch Live where
it is read, in progress_tracker.scan.  Confirmed by checking that seven
tests fail against the old target first, rather than assuming.

Mixins declare the attributes their concrete class constructs, which
basedpyright reports as uninitialised instance variables.  The rule is
disabled per-file on the nine mixin modules rather than globally, so it
keeps covering every real class.  This regression started in the
previous commit and went unnoticed because the pre-commit hook passes
on warnings; it is now back to zero.

FixOrchestrator.run and MergeProgressTracker._generate_display_text are
decomposed to come under the length limit.  The latter now takes its
twelve counters as one named snapshot, preserving the single lock
acquisition the original used to avoid a torn read on rich's refresh
thread.

Findings drop from 42 to 35; the score rises from 48 to 54.
