+++ src/dependamerge/gerrit/submit_manager.py
@@ -12,19 +12,28 @@ It supports:
 - Submit with pre-flight checks (submittable status)
 - Error handling and result tracking
 - Dry-run mode for previewing operations
+
+The submit status vocabulary and result construction live in the sibling
+module ``_submit_results``, and the per-change review and submit calls
+live in ``_submit_operations``; both are mixed back into
+``GerritSubmitManager`` here, so this module's surface is unchanged.
+``build_client`` is deliberately resolved in *this* module's namespace
+only, so that substituting it here is observed by the code that uses it.
 """
 
 from __future__ import annotations
 
 import logging
-import time
 from concurrent.futures import ThreadPoolExecutor
-from enum import Enum
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
+# The redundant ``as`` aliases below mark deliberate re-exports: every name
+# here has always been reachable as
+# ``dependamerge.gerrit.submit_manager.<name>``, even where only the sibling
+# modules still reference it.
+from dependamerge.gerrit.client import GerritAuthError as GerritAuthError
+from dependamerge.gerrit.client import GerritRestError as GerritRestError
 from dependamerge.gerrit.client import (
-    GerritAuthError,
-    GerritRestError,
     build_client,
 )
 from dependamerge.gerrit.models import (
@@ -33,6 +42,11 @@ from dependamerge.gerrit.models import (
     GerritSubmitResult,
 )
 
+# The sibling modules below carry the parts of this module that never touch
+# build_client.
+from ._submit_operations import _GerritSubmitOperationMixin
+from ._submit_results import SubmitStatus as SubmitStatus
+
 if TYPE_CHECKING:
     from dependamerge.progress_tracker import MergeProgressTracker
 
@@ -40,20 +54,7 @@ if TYPE_CHECKING:
 log = logging.getLogger("dependamerge.gerrit.submit_manager")
 
 
-class SubmitStatus(str, Enum):
-    """Status values for submit operations."""
-
-    PENDING = "pending"
-    REVIEWING = "reviewing"
-    REVIEWED = "reviewed"
-    SUBMITTING = "submitting"
-    SUBMITTED = "submitted"
-    FAILED = "failed"
-    BLOCKED = "blocked"
-    SKIPPED = "skipped"
-
-
-class GerritSubmitManager:
+class GerritSubmitManager(_GerritSubmitOperationMixin):
     """
     Manages parallel approval and submission of Gerrit changes.
 
@@ -238,309 +239,6 @@ class GerritSubmitManager:
             else:
                 tracker.merge_failure(change_key)
         return result
-
-    def _precheck_submit(
-        self,
-        change: GerritChangeInfo,
-        dry_run: bool,
-        start_time: float,
-    ) -> GerritSubmitResult | None:
-        """Return a short-circuit result, or None when submission may proceed.
-
-        Rejects changes that cannot be submitted (not open, work in
-        progress) and satisfies dry runs without any network calls.
-        """
-        if not change.is_open:
-            return GerritSubmitResult.failure_result(
-                change_number=change.number,
-                project=change.project,
-                error=f"Change is not open (status: {change.status})",
-                duration=time.time() - start_time,
-            )
-
-        if change.work_in_progress:
-            return GerritSubmitResult.failure_result(
-                change_number=change.number,
-                project=change.project,
-                error="Change is marked as Work In Progress",
-                duration=time.time() - start_time,
-            )
-
-        if dry_run:
-            log.info(
-                "[DRY RUN] Would review and submit %s #%d",
-                change.project,
-                change.number,
-            )
-            # A dry run performs no review or submit, so report the
-            # simulated success with reviewed/submitted left False.
-            # Callers gate real side effects on ``submitted`` (e.g.
-            # closing the corresponding GitHub PR after a Gerrit
-            # submit), so a dry run must never claim it submitted.
-            return GerritSubmitResult.success_result(
-                change_number=change.number,
-                project=change.project,
-                reviewed=False,
-                submitted=False,
-                duration=time.time() - start_time,
-            )
-
-        return None
-
-    def _submit_error_result(
-        self,
-        change: GerritChangeInfo,
-        exc: Exception,
-        reviewed: bool,
-        start_time: float,
-    ) -> GerritSubmitResult:
-        """Map an exception raised during submission to a failure result.
-
-        Classifies the exception, logs it at the matching level (auth
-        and REST errors are expected; anything else is logged with a
-        traceback), and returns a failure result.
-        """
-        if isinstance(exc, GerritAuthError):
-            log.error(
-                "Authentication error for %s #%d: %s",
-                change.project,
-                change.number,
-                exc,
-            )
-            message = f"Authentication error: {exc}"
-        elif isinstance(exc, GerritRestError):
-            log.error(
-                "REST error for %s #%d: %s",
-                change.project,
-                change.number,
-                exc,
-            )
-            message = f"REST error: {exc}"
-        else:
-            log.exception(
-                "Unexpected error for %s #%d: %s",
-                change.project,
-                change.number,
-                exc,
-            )
-            message = f"Unexpected error: {exc}"
-
-        return GerritSubmitResult.failure_result(
-            change_number=change.number,
-            project=change.project,
-            error=message,
-            reviewed=reviewed,
-            duration=time.time() - start_time,
-        )
-
-    def _submit_single_change(
-        self,
-        change: GerritChangeInfo,
-        review_labels: dict[str, int],
-        dry_run: bool,
-    ) -> GerritSubmitResult:
-        """
-        Submit a single change (review + submit).
-
-        Args:
-            change: The change to submit.
-            review_labels: Labels to apply.
-            dry_run: If True, simulate without making changes.
-
-        Returns:
-            GerritSubmitResult indicating success or failure.
-        """
-        start_time = time.time()
-        reviewed = False
-
-        precheck = self._precheck_submit(change, dry_run, start_time)
-        if precheck is not None:
-            return precheck
-
-        try:
-            review_success = self._review_change(change.number, review_labels)
-            if not review_success:
-                return GerritSubmitResult.failure_result(
-                    change_number=change.number,
-                    project=change.project,
-                    error="Failed to apply review",
-                    reviewed=False,
-                    duration=time.time() - start_time,
-                )
-            reviewed = True
-            log.info(
-                "Applied review to %s #%d: %s",
-                change.project,
-                change.number,
-                review_labels,
-            )
-
-            submit_success = self._submit_change(change.number)
-            if not submit_success:
-                return GerritSubmitResult.failure_result(
-                    change_number=change.number,
-                    project=change.project,
-                    error="Failed to submit (change may not be submittable)",
-                    reviewed=reviewed,
-                    duration=time.time() - start_time,
-                )
-            log.info(
-                "Submitted %s #%d",
-                change.project,
-                change.number,
-            )
-
-            return GerritSubmitResult.success_result(
-                change_number=change.number,
-                project=change.project,
-                reviewed=reviewed,
-                submitted=True,
-                duration=time.time() - start_time,
-            )
-
-        except Exception as exc:
-            return self._submit_error_result(change, exc, reviewed, start_time)
-
-    def _review_change(
-        self,
-        change_number: int,
-        labels: dict[str, int],
-    ) -> bool:
-        """
-        Apply a review (vote) to a change.
-
-        Args:
-            change_number: The change number.
-            labels: Labels to apply (e.g., {"Code-Review": 2}).
-
-        Returns:
-            True if successful, False otherwise.
-        """
-        endpoint = f"/changes/{change_number}/revisions/current/review"
-        payload = {"labels": labels}
-
-        try:
-            self._client.post(endpoint, data=payload)
-            return True
-        except GerritRestError as exc:
-            log.warning("Failed to review change %d: %s", change_number, exc)
-            return False
-
-    def _submit_change(self, change_number: int) -> bool:
-        """
-        Submit a change.
-
-        Args:
-            change_number: The change number.
-
-        Returns:
-            True if successful, False otherwise.
-        """
-        endpoint = f"/changes/{change_number}/submit"
-
-        try:
-            self._client.post(endpoint)
-            return True
-        except GerritRestError as exc:
-            log.warning("Failed to submit change %d: %s", change_number, exc)
-            return False
-
-    def review_only(
-        self,
-        changes: list[GerritChangeInfo],
-        review_labels: dict[str, int] | None = None,
-        dry_run: bool = False,
-    ) -> list[GerritSubmitResult]:
-        """
-        Apply reviews without submitting.
-
-        Useful for approving changes that need additional verification.
-
-        Args:
-            changes: List of changes to review.
-            review_labels: Labels to apply.
-            dry_run: If True, simulate without making changes.
-
-        Returns:
-            List of results indicating review success/failure.
-        """
-        if review_labels is None:
-            review_labels = {"Code-Review": 2}
-
-        results: list[GerritSubmitResult] = []
-
-        for change in changes:
-            start_time = time.time()
-
-            if dry_run:
-                log.info(
-                    "[DRY RUN] Would review %s #%d with %s",
-                    change.project,
-                    change.number,
-                    review_labels,
-                )
-                results.append(
-                    GerritSubmitResult.success_result(
-                        change_number=change.number,
-                        project=change.project,
-                        reviewed=True,
-                        submitted=False,
-                        duration=time.time() - start_time,
-                    )
-                )
-                continue
-
-            success = self._review_change(change.number, review_labels)
-            if success:
-                results.append(
-                    GerritSubmitResult.success_result(
-                        change_number=change.number,
-                        project=change.project,
-                        reviewed=True,
-                        submitted=False,
-                        duration=time.time() - start_time,
-                    )
-                )
-            else:
-                results.append(
-                    GerritSubmitResult.failure_result(
-                        change_number=change.number,
-                        project=change.project,
-                        error="Failed to apply review",
-                        duration=time.time() - start_time,
-                    )
-                )
-
-        return results
-
-    def get_submit_summary(self, results: list[GerritSubmitResult]) -> dict[str, Any]:
-        """
-        Generate a summary of submit results.
-
-        Args:
-            results: List of submit results.
-
-        Returns:
-            Dictionary with summary statistics.
-        """
-        total = len(results)
-        successful = sum(1 for r in results if r.success)
-        failed = total - successful
-        reviewed = sum(1 for r in results if r.reviewed)
-        submitted = sum(1 for r in results if r.submitted)
-        total_duration = sum(r.duration_seconds for r in results)
-
-        return {
-            "total": total,
-            "successful": successful,
-            "failed": failed,
-            "reviewed": reviewed,
-            "submitted": submitted,
-            "total_duration_seconds": round(total_duration, 2),
-            "average_duration_seconds": (
-                round(total_duration / total, 2) if total > 0 else 0.0
-            ),
-        }
 
 
 def create_submit_manager(
